+++ src/components/app/App.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Header } from '../header/Header';
+import { Header } from '../header/HeaderFromLecture';
 
 // todo: здесь нужно сделать импорт Footer из '../footer/Footer'
